refactor: use preprocesser directives to switch between different PatchAsync implementations
